# api-service/docs/swagger.yaml
basePath: /
definitions:
  models.ChangePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        maxLength: 128
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
  models.RegisterRequest:
    properties:
      email:
        maxLength: 100
        type: string
      password:
        maxLength: 128
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - email
    - password
    - username
    type: object
  models.RegisterResponse:
    properties:
      email:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  models.UpdateUserRequest:
    properties:
      email:
        maxLength: 100
        type: string
      username:
        maxLength: 50
        minLength: 3
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      last_login:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.UserPreferences:
    properties:
      email_enabled:
        type: boolean
      frequency:
        description: e.g., "immediate", "daily"
        type: string
    type: object
host: localhost
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Production-ready SaaS starter kit API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Azlo Go Boilerplate API
  version: 1.0.1
paths:
  /api/v1/admin/db-stats:
    get:
      description: Get internal database connection pool stats
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Database Statistics
      tags:
      - admin
  /api/v1/password:
    put:
      consumes:
      - application/json
      description: Verifies current password and updates to a new one
      parameters:
      - description: Password Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Current password incorrect
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Change user password
      tags:
      - profile
  /api/v1/preferences:
    get:
      description: Retrieves current logged-in user preferences
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserPreferences'
      security:
      - Bearer: []
      summary: Get user preferences
      tags:
      - preferences
    put:
      consumes:
      - application/json
      description: Allows the user to set email notification status and digest frequency
      parameters:
      - description: Preference Settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UserPreferences'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserPreferences'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update user preferences
      tags:
      - preferences
  /api/v1/profile:
    get:
      description: Retrieves detailed profile information for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
      security:
      - Bearer: []
      summary: Get current profile
      tags:
      - profile
    put:
      consumes:
      - application/json
      description: Updates username or email for the current user
      parameters:
      - description: Update Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: user_id
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Update profile info
      tags:
      - profile
  /api/v1/protected:
    get:
      description: Simple check to verify JWT authentication is working
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Test protected endpoint
      tags:
      - profile
  /api/v1/users:
    get:
      description: Get a paginated list of active users (Admin utility)
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
      security:
      - Bearer: []
      summary: List users
      tags:
      - admin
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with username, email, and password
      parameters:
      - description: Registration Info
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RegisterResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: User already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - auth
schemes:
- https
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"